Duy - Add Line Chart and Pie Chart

+++ MedicaWeb_MVC/ViewModels/Charts/PieChartVM.cs
@@ -2,7 +2,7 @@
 {
     public class PieChartVM
     {
-        public decimal[] Series { get; set; }
+        public int[] Series { get; set; }
         public string[] Labels { get; set; }
     }
 }
